refactor(step2): 기능 분리를 위해 Calculator 클래스에 기반 메소드를 이동
