+++ src/main/java/step2/Calculator.java
@@ -1,0 +1,19 @@
+package step2;
+
+public class Calculator {
+    static int add(int a, int b) {
+        return a + b;
+    }
+
+    static int subtract(int a, int b) {
+        return a - b;
+    }
+
+    static int multiple(int a, int b) {
+        return a * b;
+    }
+
+    static int divide(int a, int b) {
+        return a / b;
+    }
+}
