+++ src/main/java/step2/StringCalculator.java
@@ -2,19 +2,4 @@ package step2;
 
 public class StringCalculator {
 
-    static int add(int a, int b) {
-        return a + b;
-    }
-
-    static int subtract(int a, int b) {
-        return a - b;
-    }
-
-    static int multiple(int a, int b) {
-        return a * b;
-    }
-
-    static int divide(int a, int b) {
-        return a / b;
-    }
 }
